Add farewellInSpanish activity and apply retry options
This commit does several things: 1) it introduces a new 'farewellInSpanish' activity that is implemented in the 'GreetingActivitiesImpl' class and used in the 'GreetingWorkflowImpl' class. 2) Retry options are added in both the 'GreetingWorkflowImpl' and 'Starter' classes. This ensures activity attempts are retried under certain configurable circumstances. 3) The activity implementation registration in the 'GreetingWorker' class has also been changed.